Stop force field minimization when energy improvement slows down to improve performance (#95)
When tracking the energy after each force field minimization step, we noticed that the energy tends to 'converge' far before hitting 1000 iterations.

For many of these molecules, there is still a small energy decrease over every few iterations after converging (however the energy levels tend to oscillate). I started looking into how small changes in energy (~10-50) impact coordinates, and it does not seem to make a huge visual difference. (See PR #95 for graphs of specific examples).

In one representative example, the energy drops by around 50 between iterations 200 and 1000, however most of the significant movements occur in the first 200 iterations. After this commit, the minimization would stop once the net energy decrease in the last ITERATION_HISTORY_SIZE iterations is less than MAX_NET_ENERGY_CHANGE. I currently have these values set to 100 and 20 respectively. This change caused 99% of the minimizations in my testing to stop early (97% stop before hitting 300 iterations).

For an example set of 4184 molecules, this improved coordgen time from 24s to 10s (total), and cut the number of molecules for which coordgen took longer than 0.1s from 13 to 2.

+++ CoordgenMinimizer.h
@@ -390,7 +390,7 @@ private:
              std::vector<sketcherMinimizerInteraction*>>
         _extraInteractionsOfMolecule;
 
-    float m_maxIterations;
+    unsigned int m_maxIterations;
     float m_precision;
 };
